+++ core/src/commonMain/kotlin/ManufacturerData.kt
@@ -4,7 +4,7 @@ public class ManufacturerData(
     /**
      * Two-octet [Company Identifier Code][https://www.bluetooth.com/specifications/assigned-numbers/company-identifiers/]
      */
-    public val code: Short,
+    public val code: Int,
 
     /**
      * the Manufacturer Data (not including the leading two identifier octets)
